:recycle: refactor: Add ItemLikeController, modified request mapping change
